backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
langchain
langchain-community
langchain-google-genai
langchain-groq
faiss-cpu
tiktoken
python-dotenv
asyncpg
sqlmodel
pypdf
sentence-transformers
langchain-huggingface
